modify mochi lora
