+++ ppdiffusers/examples/inference/text_to_video_generation_mochi.py
@@ -17,7 +17,7 @@ from ppdiffusers import MochiPipeline
 from ppdiffusers.utils import export_to_video
 
 
-pipe = MochiPipeline.from_pretrained("genmo/mochi-1-preview", 
+pipe = MochiPipeline.from_pretrained("/home/lcjgrp/lizhijun/llm/mochi-1-preview-pd", 
                                      variant="bf16", 
                                      paddle_dtype=paddle.bfloat16,
                                      low_cpu_mem_usage=True, 
@@ -28,6 +28,6 @@ pipe = MochiPipeline.from_pretrained("genmo/mochi-1-preview",
 pipe.enable_vae_tiling()
 
 prompt = "Close-up of a chameleon's eye, with its scaly skin changing color. Ultra high resolution 4k."
-frames = pipe(prompt, num_frames=84).frames[0]
+frames = pipe(prompt, num_frames=30).frames[0]
 
 export_to_video(frames, "mochi.mp4", fps=30)
